requirements.txt
###### Requirements without Version Specifiers ######
speedtest-cli
python-crontab
pandas
###### Requirements with Version Specifiers ######
